requirements/dev.txt
environs==9.3.3
kafka-python==2.0.2
marshmallow==3.13.0
python-dotenv==0.19.0
